fix the xcm bridge runtime and add event
