+++ xcm-bridge/src/mock/mod.rs
@@ -3,313 +3,18 @@
 
 #![cfg(test)]
 
-use std::marker::PhantomData;
-use std::result;
-
-use frame_support::{construct_runtime, pallet_prelude::ConstU32, parameter_types, sp_runtime::{
+use frame_support::sp_runtime::{
     AccountId32,
-    BuildStorage,
-    Perbill, testing::H256, traits::{BlakeTwo256, IdentityLookup},
-}, traits::{AsEnsureOriginWithArg, ConstU128, Everything, Nothing}};
-use frame_system::{EnsureSigned, self as system};
-use polkadot_parachain_primitives::primitives::Sibling;
+    BuildStorage};
 use sp_io::TestExternalities;
-use xcm::latest::{prelude::*, Weight as XCMWeight};
-use xcm_builder::{AccountId32Aliases, AllowTopLevelPaidExecutionFrom, AllowUnpaidExecutionFrom, CurrencyAdapter, FixedWeightBounds, FungiblesAdapter, IsConcrete, NativeAsset, NoChecking, ParentIsPreset, RelayChainAsNative, SiblingParachainAsNative, SiblingParachainConvertsVia, SignedAccountId32AsNative, SovereignSignedViaLocation, TakeWeightCredit};
-use xcm_executor::{Config, traits::{Error as ExecutionError, MatchesFungibles, WithOriginFilter}, XcmExecutor};
 use xcm_simulator::{decl_test_network, decl_test_parachain, decl_test_relay_chain, TestExt};
-
-use crate as sygma_xcm_bridge;
 
 pub mod relay;
 pub mod para;
 
-type Block = frame_system::mocking::MockBlock<Runtime>;
-
-pub(crate) type Balance = u128;
-
-pub type AccountId = AccountId32;
-
-construct_runtime!(
-	pub enum Runtime {
-		System: frame_system::{Pallet, Call, Storage, Config<T>, Event<T>},
-        Assets: pallet_assets::{Pallet, Call, Storage, Event<T>},
-		Balances: pallet_balances::{Pallet, Call, Storage, Config<T>, Event<T>},
-        ParachainInfo: pallet_parachain_info::{Pallet, Storage, Config<T>},
-        CumulusXcm: cumulus_pallet_xcm::{Pallet, Event<T>, Origin},
-		SygmaXcmBridge: sygma_xcm_bridge::{Pallet, Event<T>},
-		SygmaBridgeForwarder: sygma_bridge_forwarder::{Pallet, Event<T>},
-	}
-);
-
-parameter_types! {
-	pub const BlockHashCount: u64 = 250;
-	pub const MaximumBlockLength: u32 = 2 * 1024;
-	pub const AvailableBlockRatio: Perbill = Perbill::one();
-	pub const MaxLocks: u32 = 100;
-	pub const MinimumPeriod: u64 = 1;
-}
-
-parameter_types! {
-	pub const XcmBridgePalletIndex: u8 = 6;
-	pub RegisteredExtrinsics: Vec<(u8, Vec<u8>)> = [
-		(XcmBridgePalletIndex::get(), b"transfer".to_vec()),
-	].to_vec();
-}
-
-impl frame_system::Config for Runtime {
-    type BaseCallFilter = frame_support::traits::Everything;
-    type Block = Block;
-    type BlockWeights = ();
-    type BlockLength = ();
-    type RuntimeOrigin = RuntimeOrigin;
-    type RuntimeCall = RuntimeCall;
-    type Nonce = u64;
-    type Hash = H256;
-    type Hashing = BlakeTwo256;
-    type AccountId = AccountId32;
-    type Lookup = IdentityLookup<Self::AccountId>;
-    type RuntimeEvent = RuntimeEvent;
-    type BlockHashCount = BlockHashCount;
-    type DbWeight = ();
-    type Version = ();
-    type PalletInfo = PalletInfo;
-    type AccountData = pallet_balances::AccountData<Balance>;
-    type OnNewAccount = ();
-    type OnKilledAccount = ();
-    type SystemWeightInfo = ();
-    type SS58Prefix = ();
-    type OnSetCode = ();
-    type MaxConsumers = ConstU32<2>;
-}
-
-parameter_types! {
-	pub const ExistentialDeposit: Balance = 1;
-}
-
-impl pallet_balances::Config for Runtime {
-    type Balance = Balance;
-    type DustRemoval = ();
-    type RuntimeEvent = RuntimeEvent;
-    type ExistentialDeposit = ExistentialDeposit;
-    type AccountStore = System;
-    type WeightInfo = ();
-    type MaxLocks = ();
-    type MaxReserves = ();
-    type ReserveIdentifier = [u8; 8];
-    type FreezeIdentifier = ();
-    type MaxFreezes = ();
-    type RuntimeHoldReason = ();
-    type MaxHolds = ();
-}
-
-parameter_types! {
-	pub const AssetDeposit: Balance = 1; // 1 Unit deposit to create asset
-	pub const ApprovalDeposit: Balance = 1;
-	pub const AssetsStringLimit: u32 = 50;
-	pub const MetadataDepositBase: Balance = 1;
-	pub const MetadataDepositPerByte: Balance = 1;
-}
-
-pub type AssetId = u32;
-
-impl pallet_assets::Config for Runtime {
-    type RuntimeEvent = RuntimeEvent;
-    type Balance = Balance;
-    type AssetId = AssetId;
-    type AssetIdParameter = codec::Compact<u32>;
-    type Currency = Balances;
-    type CreateOrigin = AsEnsureOriginWithArg<EnsureSigned<AccountId32>>;
-    type ForceOrigin = frame_system::EnsureRoot<Self::AccountId>;
-    type AssetDeposit = AssetDeposit;
-    type AssetAccountDeposit = ConstU128<10>;
-    type MetadataDepositBase = MetadataDepositBase;
-    type MetadataDepositPerByte = MetadataDepositPerByte;
-    type ApprovalDeposit = ApprovalDeposit;
-    type StringLimit = AssetsStringLimit;
-    type RemoveItemsLimit = ConstU32<1000>;
-    type Freezer = ();
-    type Extra = ();
-    type CallbackHandle = ();
-    type WeightInfo = pallet_assets::weights::SubstrateWeight<Runtime>;
-    #[cfg(feature = "runtime-benchmarks")]
-    type BenchmarkHelper = ();
-}
-
-impl sygma_xcm_bridge::Config for Runtime {
-    type RuntimeEvent = RuntimeEvent;
-    type PalletIndex = XcmBridgePalletIndex;
-
-    type Weigher = FixedWeightBounds<UnitWeightCost, RuntimeCall, MaxInstructions>;
-    type XcmExecutor = XcmExecutor<XcmConfig>;
-    type SelfLocation = SelfLocation;
-}
-
-impl sygma_bridge_forwarder::Config for Runtime {
-    type RuntimeEvent = RuntimeEvent;
-    type SygmaBridge = ();
-    type XCMBridge = ();
-}
-
-impl pallet_parachain_info::Config for Runtime {}
-
-pub struct XcmConfig;
-
-impl Config for XcmConfig {
-    type RuntimeCall = RuntimeCall;
-    type XcmSender = XcmRouter;
-    type AssetTransactor = (CurrencyTransactor, FungiblesTransactor);
-    type OriginConverter = XcmOriginToTransactDispatchOrigin;
-    type IsReserve = NativeAsset;
-    type IsTeleporter = ();
-    type UniversalLocation = SelfLocation;
-    type Barrier = Barrier;
-    type Weigher = FixedWeightBounds<UnitWeightCost, RuntimeCall, MaxInstructions>;
-    type Trader = DynamicWeightTrader<
-        WeightPerSecond,
-        <Runtime as pallet_assets::Config>::AssetId,
-        AssetsRegistry,
-        helper::XTransferTakeRevenue<Self::AssetTransactor, AccountId, ParaTreasuryAccount>,
-    >;
-    type ResponseHandler = ();
-    type AssetTrap = ();
-    type AssetClaims = ();
-    type SubscriptionService = ();
-    type PalletInstancesInfo = AllPalletsWithSystem;
-    type MaxAssetsIntoHolding = ConstU32<64>;
-    type AssetLocker = ();
-    type AssetExchanger = ();
-    type FeeManager = ();
-    type MessageExporter = ();
-    type UniversalAliases = Nothing;
-    type CallDispatcher = WithOriginFilter<Everything>;
-    type SafeCallFilter = Everything;
-    type Aliasers = ();
-}
-
-// TODO: add xcm simulator
-pub type XcmRouter = ParachainXcmRouter<ParachainInfo>;
-
-pub type Barrier = (
-    TakeWeightCredit,
-    AllowTopLevelPaidExecutionFrom<Everything>,
-    AllowUnpaidExecutionFrom<Everything>,
-);
-
-pub type LocationToAccountId = (
-    ParentIsPreset<AccountId>,
-    SiblingParachainConvertsVia<Sibling, AccountId>,
-    AccountId32Aliases<RelayNetwork, AccountId>,
-);
-
-pub type XcmOriginToTransactDispatchOrigin = (
-    SovereignSignedViaLocation<LocationToAccountId, RuntimeOrigin>,
-    RelayChainAsNative<RelayChainOrigin, RuntimeOrigin>,
-    SiblingParachainAsNative<cumulus_pallet_xcm::Origin, RuntimeOrigin>,
-    SignedAccountId32AsNative<RelayNetwork, RuntimeOrigin>,
-);
-
-parameter_types! {
-    pub const RelayNetwork: NetworkId = NetworkId::Rococo;
-    pub RelayChainOrigin: RuntimeOrigin = cumulus_pallet_xcm::Origin::Relay.into();
-	pub UnitWeightCost: XCMWeight = 1u64.into();
-	pub const MaxInstructions: u32 = 100;
-}
-
-impl cumulus_pallet_xcm::Config for Runtime {
-    type RuntimeEvent = RuntimeEvent;
-    type XcmExecutor = XcmExecutor<XcmConfig>;
-}
-
-parameter_types! {
-    pub NativeLocation: MultiLocation = MultiLocation::here();
-    pub UsdtAssetId: AssetId = 0;
-	pub UsdtLocation: MultiLocation = MultiLocation::new(
-		1,
-		X3(
-			Parachain(2005),
-			slice_to_generalkey(b"sygma"),
-			slice_to_generalkey(b"usdt"),
-		),
-	);
-    pub CheckingAccount: AccountId32 = AccountId32::new([102u8; 32]);
-}
-
-pub struct SimpleForeignAssetConverter(PhantomData<()>);
-
-impl MatchesFungibles<AssetId, Balance> for SimpleForeignAssetConverter {
-    fn matches_fungibles(a: &MultiAsset) -> result::Result<(AssetId, Balance), ExecutionError> {
-        match (&a.fun, &a.id) {
-            (Fungible(ref amount), Concrete(ref id)) => {
-                if id == &UsdtLocation::get() {
-                    Ok((UsdtAssetId::get(), *amount))
-                } else {
-                    Err(ExecutionError::AssetNotHandled)
-                }
-            }
-            _ => Err(ExecutionError::AssetNotHandled),
-        }
-    }
-}
-
-pub type CurrencyTransactor = CurrencyAdapter<
-    // Use this currency:
-    Balances,
-    // Use this currency when it is a fungible asset matching the given location or name:
-    IsConcrete<NativeLocation>,
-    // Convert an XCM MultiLocation into a local account id:
-    LocationToAccountId,
-    // Our chain's account ID type (we can't get away without mentioning it explicitly):
-    AccountId32,
-    // We don't track any teleports of `Balances`.
-    (),
->;
-
-/// Means for transacting assets besides the native currency on this chain.
-pub type FungiblesTransactor = FungiblesAdapter<
-    // Use this fungibles implementation:
-    Assets,
-    // Use this currency when it is a fungible asset matching the given location or name:
-    SimpleForeignAssetConverter,
-    // Convert an XCM MultiLocation into a local account id:
-    LocationToAccountId,
-    // Our chain's account ID type (we can't get away without mentioning it explicitly):
-    AccountId32,
-    // Disable teleport.
-    NoChecking,
-    // The account to use for tracking teleports.
-    CheckingAccount,
->;
-
-parameter_types! {
-	pub SelfLocation: MultiLocation = MultiLocation::new(1, X1(Parachain(ParachainInfo::parachain_id().into())));
-}
-
-pub fn new_test_ext() -> sp_io::TestExternalities {
-    let t = frame_system::GenesisConfig::<Runtime>::default().build_storage().unwrap();
-    let mut ext = sp_io::TestExternalities::new(t);
-    ext.execute_with(|| System::set_block_number(1));
-    ext
-}
-
-// Checks events against the latest. A contiguous set of events must be provided. They must
-// include the most recent event, but do not have to include every past event.
-pub fn assert_events(mut expected: Vec<RuntimeEvent>) {
-    let mut actual: Vec<RuntimeEvent> =
-        system::Pallet::<Runtime>::events().iter().map(|e| e.event.clone()).collect();
-
-    expected.reverse();
-
-    for evt in expected {
-        let next = actual.pop().expect("event expected");
-        assert_eq!(next, evt, "Events don't match");
-    }
-}
-
 pub const ALICE: AccountId32 = AccountId32::new([0u8; 32]);
 pub const BOB: AccountId32 = AccountId32::new([1u8; 32]);
 pub const ENDOWED_BALANCE: u128 = 100_000_000;
-pub const TEST_THRESHOLD: u32 = 2;
 
 pub type ParaBalances = pallet_balances::Pallet<para::Runtime>;
 pub type ParaAssets = pallet_assets::Pallet<para::Runtime>;
@@ -405,17 +110,4 @@ pub fn relay_ext() -> sp_io::TestExternalities {
     let mut ext = sp_io::TestExternalities::new(t);
     ext.execute_with(|| System::set_block_number(1));
     ext
-}
-
-pub fn slice_to_generalkey(key: &[u8]) -> Junction {
-    let len = key.len();
-    assert!(len <= 32);
-    GeneralKey {
-        length: len as u8,
-        data: {
-            let mut data = [0u8; 32];
-            data[..len].copy_from_slice(key);
-            data
-        },
-    }
 }
